PRSD-976: Dependency inject view model factories (#633)
